requirements.txt
django
Django-import-export
Django-allauth
Django-filter
djangorestframework
six
pillow
django-allauth[socialaccount]
django-htmx
django-cleanup
django-channels
django-channels[daphne]

